Destroy Associations in Background Job
Motivation:
  - Sometimes cascading association deletions can cause timeouts due to
  IO issue. Perhaps a model has associations that are destroyed on
  deletion which in turn trigger other deletions and this can continue
  down a complex tree. Along this tree you may also hit other IO
  operations. Such deep deletions can lead to server timeouts while
  awaiting completion and really the user may not notice all the
  changes on their side immediately making them wait unnecesarially or
  worse causing a timeout during the operation.

  It is possible to move the association deletion into jobs
  today, however it is pretty boilerplate and it would be nice to make
  it easier to do.

  This pr builds off the work of @georgeclaghorn in PR #34901 and is
  currently a simple proposal of how this could work. I branched off his
  pr for now to quickly product this proof of concept.

Related Issues:
  - #34901
  - rails/actionmailbox#16

Changes:
  - Added a new job to handle deleting associations.
  - Added a new `dependent` option on `has_many` associations,
  `destroy_async` (Open to other names).
  - Added a single proof of concept test.

+++ activerecord/lib/active_record/associations/belongs_to_association.rb
@@ -11,6 +11,16 @@ module ActiveRecord
         when :destroy
           target.destroy
           raise ActiveRecord::Rollback unless target.destroyed?
+        when :destroy_later
+          id = owner.send(reflection.foreign_key.to_sym)
+          primary_key_column = reflection.
+              active_record_primary_key.to_sym
+          enqueue_destroy_association(owner_model_name: owner.class.to_s,
+                          owner_id: owner.id,
+                          assoc_class: reflection.klass.to_s,
+                          assoc_ids: [id],
+                          assoc_primary_key_column: primary_key_column,
+                         owner_ensuring_destroy_method: options.fetch(:owner_ensuring_destroy, nil))
         else
           target.send(options[:dependent])
         end
